Comm/old/requirement.txt
django 
tensorflow 
pandas 
mysqlclient 
djangorestframework
scikit-learn
matplotlib
django-cors-headers